[Update] Floating-Point Columns - Replaced `float()` with `double()`, since `float()` now creates a `float(53)` type column as opposed to `double` - Replaced the removed `unsignedDecimal` with `decimal` as the "unsigned" modifier has been deprecated by MySQL

+++ database/migrations/2018_08_22_154146_create_media_ratings_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('user_id');
             $table->morphs('model');
-            $table->float('rating')->default(MediaRating::MAX_RATING_VALUE);
+            $table->double('rating')->default(MediaRating::MAX_RATING_VALUE);
             $table->text('description')->nullable();
             $table->text('note')->nullable();
             $table->timestamps();
